test(industrial): fold dedicated-bounds tests into existing module
CI quality-guardrail failure: the new `mod` include grew
kirra_verifier_service.rs from 1831 -> 1835 lines, past its guardrail ceiling
(the L3 decomposition intentionally shrank that file). Rather than spend that
file's ceiling on test-module registration, fold the two dedicated CANopen/CIP
bound-wiring tests into the existing industrial_replay_handler_tests.rs (reusing
its svc() harness — the separate file had duplicated it) and drop the separate
file + include. Behaviour and coverage identical; god file back to 1831.

cargo test --bin kirra_verifier_service industrial: 5 passed. fmt + clippy clean.
Guardrails satisfied.
